write a

+++ files/programs/kernel/graphics/graphics.cpp
@@ -12,7 +12,7 @@ namespace {
     }
 }
 
-void FrameBufferWriter::WritePixel(int x, int y, uint8_t color) {
+void FrameBufferWriter::WritePixel(int x, int y, unsigned int color) {
     uint8_t* pixel = &frame_buffer_.base_address[GetPixelBytes(frame_buffer_.pixel_format) * (frame_buffer_.line * y + x)];
     pixel[0] = color >> 16;
     pixel[1] = (color >> 8) & 0xff;
