+++ files/programs/kernel/graphics/graphics.hpp
@@ -5,7 +5,7 @@
 class FrameBufferWriter {
     public:
         FrameBufferWriter(const FrameBuffer& frame_buffer): frame_buffer_{frame_buffer} {}
-        void WritePixel(int x, int y, uint8_t color);
+        void WritePixel(int x, int y, unsigned int color);
 
     private:
         const FrameBuffer& frame_buffer_;
